#236 refactor : double -> float 로 단위 변경

+++ src/main/java/com/floney/floney/analyze/dto/response/AnalyzeResponseByAsset.java
@@ -8,19 +8,19 @@ import java.util.Map;
 
 @Getter
 public class AnalyzeResponseByAsset {
-    private final double difference;
-    private final double initAsset;
-    private final double currentAsset;
+    private final float difference;
+    private final float initAsset;
+    private final float currentAsset;
     private final Map<LocalDate, AssetInfo> assetInfo;
 
-    private AnalyzeResponseByAsset(double difference, double initAsset, double currentAsset, Map<LocalDate, AssetInfo> assetInfo) {
+    private AnalyzeResponseByAsset(float difference, float initAsset, float currentAsset, Map<LocalDate, AssetInfo> assetInfo) {
         this.difference = difference;
         this.currentAsset = currentAsset;
         this.initAsset = initAsset;
         this.assetInfo = assetInfo;
     }
 
-    public static AnalyzeResponseByAsset of(double difference, double initAsset, double currentAsset, Map<LocalDate,AssetInfo> assetInfo) {
+    public static AnalyzeResponseByAsset of(float difference, float initAsset, float currentAsset, Map<LocalDate,AssetInfo> assetInfo) {
         return new AnalyzeResponseByAsset(difference, initAsset, currentAsset, assetInfo);
     }
 }
